MODIFIED ASSIGNMENT AS REQUESTED
I made changes to portion 1 and 2 as instructed.
For portion 1:
provided more clarification on the sampling procedure and showing it on the python code line by line.
For portion 2:
looked at R code of the blog and noticed that both approaches are using same sampling methodologies and producing same results.